refactor(bulk-upload): De-dupe S3 functions
