+++ server/routes/establishments/bulkUpload/validate.js
@@ -120,70 +120,7 @@ const restoreExistingEntities = async (
   }
 };
 
-const saveResponse = async (req, res, statusCode, body, headers) => {
-  if (!Number.isInteger(statusCode) || statusCode < 100) {
-    statusCode = 500;
-  }
-
-  return s3
-    .putObject({
-      Bucket,
-      Key: `${req.establishmentId}/intermediary/${req.buRequestId}.json`,
-      Body: JSON.stringify({
-        url: req.url,
-        startTime: req.startTime,
-        endTime: new Date().toISOString(),
-        responseCode: statusCode,
-        responseBody: body,
-        responseHeaders: typeof headers === 'object' ? headers : undefined,
-      }),
-    })
-    .promise();
-};
-
-const uploadAsJSON = async (username, establishmentId, content, key) => {
-  try {
-    await s3
-      .putObject({
-        Bucket,
-        Key: key,
-        Body: JSON.stringify(content, null, 2),
-        ContentType: 'application/json',
-        Metadata: {
-          username,
-          establishmentId: String(establishmentId),
-        },
-      })
-      .promise();
-  } catch (err) {
-    console.error('uploadAsJSON: ', err);
-    throw new Error(`Failed to upload S3 object: ${key}`);
-  }
-};
-
-const downloadContent = async (key, size, lastModified) => {
-  try {
-    const filenameRegex = /^(.+\/)*(.+)\.(.+)$/;
-
-    return await s3
-      .getObject({
-        Bucket,
-        Key: key,
-      })
-      .promise()
-      .then((objData) => ({
-        key,
-        data: objData.Body.toString(),
-        filename: key.match(filenameRegex)[2] + '.' + key.match(filenameRegex)[3],
-        username: objData.Metadata.username,
-        size,
-        lastModified,
-      }));
-  } catch (err) {
-    console.error(`api/establishment/bulkupload/downloadFile: ${key})\n`, err);
-    throw new Error(`Failed to download S3 object: ${key}`);
-  }
-};
+const { saveResponse, uploadAsJSON, downloadContent } = require('./s3');
 
 const loadWorkerQualifications = async (
   lineValidator,
